feat(storage): integrate Async Storage for react-native
Integrate the latest version of Async Storage into the react-native project to facilitate
persistent key-value storage. This includes the addition of the relevant dependency in
the package.json and yarn.lock files, as well as the implementation of utility functions
for saving, fetching, and removing items from the storage in the src/utils/storage.ts file.